Move node and event observer helpers to e2e/common
